test(ios): run the Liquid Glass assertions and correct them
Both Liquid Glass tests open with `guard #available(iOS 26.0, *)`, and
test-funput-kit.sh picks the first available iOS simulator — 18.6 — so
they returned early and passed empty. They had never executed once.

Running them on an iOS 26 simulator showed two assertions were wrong:

- `isInteractive` was asserted true, but 79a7337 set it false two days
  after the test was written. False is correct: the keycap runs its own
  press animation, and interactive glass would answer the same touch a
  second time.
- `backgroundColor == .clear` was never true; no commit has ever set it.
  Now asserted by alpha, which holds whether the value is nil or clear.

A third failure was ours: `vniSurface` checks that every key has a usable
frame, but its helper sweeps up every UIControl in the tree — including
the clipboard chip's paste control, which has no frame while the chip is
hidden. Scoped to keycaps.

Split into their own file, noting they need an iOS 26 simulator to mean
anything.
