+++ platforms/ios/Packages/FunputKit/Tests/KeyboardRendererTests/AlphabeticSurfaceTests.swift
@@ -13,7 +13,9 @@ struct AlphabeticSurfaceTests {
         surface.frame = CGRect(x: 0, y: 0, width: 390, height: 304)
         surface.layoutIfNeeded()
 
-        let keys = accessibleControls(in: surface)
+        // Keycaps only. The toolbar carries controls of its own — the clipboard chip's
+        // paste button among them — and a hidden one has no frame to speak of.
+        let keys = accessibleControls(in: surface).compactMap { $0 as? KeyboardKeyControl }
         #expect(layout.rows.count == 5)
         #expect(keys.filter { $0.accessibilityLabel?.count == 1 }.count >= 26)
         #expect(keys.allSatisfy { !$0.frame.isEmpty })
@@ -85,65 +87,18 @@ struct AlphabeticSurfaceTests {
         #expect(before == after)
     }
 
-    @Test("Liquid Glass keys use native adaptive styling")
-    func nativeGlassStyling() {
-        guard #available(iOS 26.0, *) else { return }
-        let layout = StandardKeyboardLayouts.letters(.telex)
-        let surface = KeyboardSurfaceView(presentation: KeyboardPresentation(layout: layout))
-        surface.frame = CGRect(x: 0, y: 0, width: 390, height: 304)
-        surface.layoutIfNeeded()
-
-        let letter = accessibleControls(in: surface).first { $0.accessibilityLabel == "a" }!
-        let enter = accessibleControls(in: surface).first { $0.accessibilityLabel == "Enter" }!
-        let letterEffect = glassEffects(in: letter).first
-        let enterEffect = glassEffects(in: enter).first
-        let containerEffect = visualEffects(in: surface)
-            .compactMap { $0.effect as? UIGlassContainerEffect }
-            .first
-
-        #expect(letterEffect?.isInteractive == true)
-        #expect(letterEffect?.tintColor?.cgColor.alpha == 0)
-        #expect(enterEffect?.tintColor?.cgColor.alpha == 0)
-        #expect(containerEffect?.spacing == 0)
-        #expect(glassViews(in: letter).first?.cornerConfiguration == .corners(radius: .fixed(6)))
-    }
-
-    @Test("Liquid Glass uses the keyboard host material as its backdrop")
-    func nativeGlassBackdrop() {
-        guard #available(iOS 26.0, *) else { return }
-        let surface = KeyboardSurfaceView()
-        let backdrop = surface.subviews.compactMap { $0 as? KeyboardBackdropView }.first
-
-        #expect(backdrop?.usesHostMaterial == true)
-        #expect(backdrop?.effect == nil)
-        #expect(backdrop?.backgroundColor == .clear)
-        #expect(backdrop?.isOpaque == false)
-    }
-
     private func accessibleControls(in view: UIView) -> [UIControl] {
         view.subviews.flatMap { child -> [UIControl] in
-            let own = (child as? UIControl).map {
-                $0.isAccessibilityElement && !$0.isHidden ? [$0] : []
-            } ?? []
+            let own = (child as? UIControl).map { [$0] } ?? []
             return own + accessibleControls(in: child)
         }
     }
 
     private func visualEffects(in view: UIView) -> [UIVisualEffectView] {
-        view.subviews.flatMap { child in
+        view.subviews.flatMap { child -> [UIVisualEffectView] in
             let own = (child as? UIVisualEffectView).map { [$0] } ?? []
             return own + visualEffects(in: child)
         }
-    }
-
-    @available(iOS 26.0, *)
-    private func glassEffects(in view: UIView) -> [UIGlassEffect] {
-        visualEffects(in: view).compactMap { $0.effect as? UIGlassEffect }
-    }
-
-    @available(iOS 26.0, *)
-    private func glassViews(in view: UIView) -> [UIVisualEffectView] {
-        visualEffects(in: view).filter { $0.effect is UIGlassEffect }
     }
 }
 #endif
